requirements.txt
Flask
gunicorn
python-binance
line-bot-sdk==1.20.0